+++ src/imio/events/core/tests/test_utils.py
@@ -163,18 +163,18 @@ class TestAgenda(unittest.TestCase):
         self.assertEqual(expanded_events[-1]["end"], "2022-12-11T13:00:00+00:00")
         events = [
             {
-                "start": "2022-11-13T12:00:00+00:00",
-                "end": "2022-11-13T12:00:00+00:00",
-                "first_start": "2022-11-13T12:00:00+00:00",
-                "first_end": "2022-11-13T12:00:00+00:00",
+                "start": "2022-11-12T23:00:00+00:00",
+                "end": "2022-11-12T23:00:00+00:00",
+                "first_start": "2022-11-12T23:00:00+00:00",
+                "first_end": "2022-11-12T23:00:00+00:00",
                 "recurrence": "RRULE:FREQ=WEEKLY;COUNT=5",
                 "open_end": False,
                 "whole_day": True,
             }
         ]
         expanded_events = expand_occurences(events)
-        self.assertEqual(expanded_events[-1]["start"], "2022-12-11T12:00:00+00:00")
-        self.assertEqual(expanded_events[-1]["end"], "2022-12-11T12:00:00+00:00")
+        self.assertEqual(expanded_events[-1]["start"], "2022-12-10T23:00:00+00:00")
+        self.assertEqual(expanded_events[-1]["end"], "2022-12-11T22:59:59+00:00")
         events = [
             {
                 "start": "2022-11-13T00:00:00+00:00",